recovery: split `wal.rs` into `recovery` mod
